app/requirements.txt
fastapi==0.110.3
python-jose==3.3.0
passlib==1.7.4
pymongo==4.7.2
python-multipart==0.0.9
uvicorn==0.29.0
